media_understanding/requirements.txt
# AgentBus 媒体理解系统依赖

# 核心异步框架
asyncio-mqtt>=0.11.0

# HTTP请求和网络
aiohttp>=3.8.0
requests>=2.28.0

# 图像处理
Pillow>=9.0.0
python-magic>=0.4.27

# 文档处理
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
openpyxl>=3.0.10
pandas>=1.5.0

# 音频处理
SpeechRecognition>=3.10.0
pydub>=0.25.0

# 文本处理
beautifulsoup4>=4.11.0
lxml>=4.9.0

# OCR支持（可选）
pytesseract>=0.3.10

# 科学计算（用于图像分析）
numpy>=1.21.0

# 配置文件处理
pyyaml>=6.0
toml>=0.10.0

# 日志和调试
loguru>=0.6.0

# 类型检查和代码质量
mypy>=1.0.0
black>=22.0.0
flake8>=5.0.0

# 测试框架
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# 开发工具
jupyter>=1.0.0
ipython>=8.0.0

# 云服务SDK（可选）
# google-generativeai>=0.3.0  # Google AI
# openai>=0.28.0              # OpenAI
# anthropic>=0.3.0            # Anthropic
# boto3>=1.26.0               # AWS服务

# 视频处理（可选）
# moviepy>=1.0.3
# ffmpeg-python>=0.2.0

# 数据库支持（可选）
# sqlalchemy>=1.4.0
# asyncpg>=0.27.0
# aiomysql>=0.1.0

# 缓存支持（可选）
# redis>=4.3.0
# aiocache>=0.12.0

# 监控和指标（可选）
# prometheus-client>=0.16.0
# structlog>=22.0.0